Remove code for lower bounding the amount of ebits
The code had a bug: the variable N (L26) should be initialized explicitly as a hermitian complex matrix (N = sdpvar(d^2,d^2,4,'hermitian','complex')). When doing so, the problem trivializes, since PPT includes LOCC, which allows for perfect teleportation using just one ebit.

The approach that would be necessary for this problem would go along the lines of arXiv:2501.17394